requirements.txt
requests==2.32.5
pandas==2.3.2
scikit-learn==1.7.2
ortools==9.14.6206
folium==0.20.0
geopy==2.4.1
python_dotenv==1.1.1
ics==0.7.2
